Add `Format' support for search()
Previously when doing a search() the format was always `l' (all
possible fields were returned). It can be often handy to limit the
fields returned and RT permits to also have `i' and `s' format
search.

Add a `Format' argument to search() (by default `l') in order to
permit to inject the corresponding `format' field in the search
HTTP request.