WIP Functional tests for authentication on REST2
Before implementing authentication in the rest2 file, I needed to verify
what authentication methods were working and how.

It turns out that I couldn't get HTTP Basic Auth to work even though the
API documents that it is one of the recommended methods for
authentication.

WIP: I have not yet had access to an RT instance that has token
authentication enabled, so the test for token-based auth does not
necessarily reflect a working basic implementation.