migrate REST2 tests to pytest
The initial implementation of tests was written with the "unittest"
library since it was what was already used for REST 1.x tests. In issue
34 sim0nx expressed a preference for using py.test, so let's migrate
over to this testing library before we have too much tests written.

pytest is able to run the tests previously written for the REST 1.x
imlementation since they were using unittest.

Since nose is unable to run tests written for pytest, we also need to
change our CI (and thus dev requirements) to run tests with pytest.
This must also include a change in coverage report method to pytest-cov,
which produces a ".coverage" sqlite database that we must now add to
ignored files.